Trait sui_benchmark::workloads::payload::Payload

pub trait Payload: Send + Sync + Debug + Display {
    // Required methods
    fn make_new_payload(&mut self, effects: &ExecutionEffects);
    fn make_transaction(&mut self) -> VerifiedTransaction;
}

A Payload is a transaction wrapper of a particular type (transfer object, shared counter, etc). Calling make_transaction() on a payload produces the transaction it is wrapping. Once that transaction is returned with effects (by quorum driver), a new payload can be generated with that effect by invoking make_new_payload(effects)

Required Methods

fn make_new_payload(&mut self, effects: &ExecutionEffects)

fn make_transaction(&mut self) -> VerifiedTransaction