Trait sui_benchmark::workloads::workload::Workload

pub trait Workload<T: Payload + ?Sized>: Send + Sync + Debug {
    // Required methods
    fn init<'life0, 'async_trait>(
        &'life0 mut self,
        proxy: Arc<dyn ValidatorProxy + Sync + Send>,
        system_state_observer: Arc<SystemStateObserver>
    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn make_test_payloads<'life0, 'async_trait>(
        &'life0 self,
        proxy: Arc<dyn ValidatorProxy + Sync + Send>,
        system_state_observer: Arc<SystemStateObserver>
    ) -> Pin<Box<dyn Future<Output = Vec<Box<T>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
}

A Workload is used to generate multiple payloads during setup phase with make_test_payloads() which are added to a local queue. We execute transactions (the queue is drained based on the target qps i.e. for 100 tps, the queue will be popped 100 times every second) with those payloads and generate new payloads (which are enqueued back to the queue) with the returned effects. The total number of payloads to generate depends on how much transaction throughput we want and the maximum number of transactions we want to have in flight. For instance, for a 100 target_qps and in_flight_ratio of 5, a maximum of 500 transactions is expected to be in flight and that many payloads are created.

Required Methods

fn init<'life0, 'async_trait>( &'life0 mut self, proxy: Arc<dyn ValidatorProxy + Sync + Send>, system_state_observer: Arc<SystemStateObserver> ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

fn make_test_payloads<'life0, 'async_trait>( &'life0 self, proxy: Arc<dyn ValidatorProxy + Sync + Send>, system_state_observer: Arc<SystemStateObserver> ) -> Pin<Box<dyn Future<Output = Vec<Box<T>>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,